move WodsId to components
